double型に変更

+++ include/game.h
@@ -6,7 +6,7 @@
 /*   By: ootsuboyoshiyuki <ootsuboyoshiyuki@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 14:34:31 by ootsuboyosh       #+#    #+#             */
-/*   Updated: 2024/11/22 17:59:52 by ootsuboyosh      ###   ########.fr       */
+/*   Updated: 2024/11/25 18:39:46 by ootsuboyosh      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,18 +34,19 @@
 // 移動方向を示すオフセット
 typedef struct s_offset
 {
-	int	dx;
-	int	dy;
-}		t_offset;
+	double	dx;
+	double	dy;
+}			t_offset;
 
-void	set_player_position(t_map *map, t_player *player);
-void	move_player(t_map *map, t_player *player, t_offset offset);
-int		handle_keypress(int keycode, t_game *game);
-void	draw_vertical_line(t_game *game, int x, int start, int end, int color);
-void	render_scene(t_game *game);
-void	draw_wall(t_game *game, t_ray *ray, int x, int draw_start,
-			int draw_end);
-void	draw_floor_and_ceiling(t_game *game);
-int		exit_program(t_game *game);
+void		set_player_position(t_map *map, t_player *player);
+void		move_player(t_map *map, t_player *player, t_offset offset);
+int			handle_keypress(int keycode, t_game *game);
+void		draw_vertical_line(t_game *game, int x, int start, int end,
+				int color);
+void		render_scene(t_game *game);
+void		draw_wall(t_game *game, t_ray *ray, int x, int draw_start,
+				int draw_end);
+void		draw_floor_and_ceiling(t_game *game);
+int			exit_program(t_game *game);
 
 #endif
